+++ Finjector.Web/ClientApp/src/components/Shared/LoadingAndErrors/ChartDebugInfo.tsx
@@ -1,4 +1,4 @@
-import { ChartData, ChartStringAndAeDetails, ChartType } from "../../types";
+import { ChartData, ChartStringAndAeDetails, ChartType } from "../../../types";
 
 interface Props {
   chartData?: ChartData;
