+++ Finjector.Web/ClientApp/src/pages/ChartStringRedirector.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Navigate, useParams } from "react-router-dom";
 import { useGetChart } from "../queries/storedChartQueries";
-import FinLoader from "../components/Shared/FinLoader";
+import FinLoader from "../components/Shared/LoadingAndErrors/FinLoader";
 
 const ChartStringRedirector: React.FC = () => {
   const { type, id } = useParams<{ type: string; id: string }>();
