+++ Finjector.Web/ClientApp/src/pages/Entry.tsx
@@ -3,7 +3,7 @@ import { Navigate, useParams, useSearchParams } from "react-router-dom";
 import ChartTypeSelector from "../components/Entry/ChartTypeSelector";
 import GlEntry from "../components/Entry/GlEntry";
 import PpmEntry from "../components/Entry/PpmEntry";
-import FinLoader from "../components/Shared/FinLoader";
+import FinLoader from "../components/Shared/LoadingAndErrors/FinLoader";
 
 import { Coa, ChartData, ChartType, SegmentData } from "../types";
 
@@ -29,8 +29,8 @@ import {
   mapSegmentQueryData,
 } from "../util/segmentMapping";
 import EntryEditButtons from "../components/Entry/EntryEditButtons";
-import { ChartDebugInfo } from "../components/Shared/ChartDebugInfo";
-import { ChartLoadingError } from "../components/Shared/ChartLoadingError";
+import { ChartDebugInfo } from "../components/Shared/LoadingAndErrors/ChartDebugInfo";
+import { ChartLoadingError } from "../components/Shared/LoadingAndErrors/ChartLoadingError";
 import FolderSearch from "../components/Entry/FolderSearch";
 import PageTitle from "../components/Shared/StyledComponents/PageTitle";
